pie chart angepasst 2.0  [#5576]

+++ src/main/java/at/ac/tuwien/sepm/ss16/qse18/gui/exam/exercise/ShowResultController.java
@@ -50,8 +50,8 @@ import java.util.Map;
         try {
             result = exerciseExamService.gradeExam(exerciseExam);
             labelGrade.setText("Grading: " + result[2]);
-            pieChartData.add(new PieChart.Data("correct " + Double.parseDouble(result[0]), Double.parseDouble(result[0])));
-            pieChartData.add(new PieChart.Data("false " + Double.parseDouble(result[1]), Double.parseDouble(result[1])));
+            pieChartData.add(new PieChart.Data("correct " + Integer.parseInt(result[0]), Integer.parseInt(result[0])));
+            pieChartData.add(new PieChart.Data("false " + Integer.parseInt(result[1]), Integer.parseInt(result[1])));
             pieChart.setTitle("correct/false answers");
             pieChart.setTitleSide(Side.BOTTOM);
             pieChart.setData(pieChartData);
@@ -71,7 +71,7 @@ import java.util.Map;
             series1.setName("false");
 
             for(Map.Entry<Topic, String[]> m: topicResult.entrySet()){
-                XYChart.Data data = new XYChart.Data(m.getKey().getTopic(), Double.parseDouble(m.getValue()[1]));
+                XYChart.Data data = new XYChart.Data(m.getKey().getTopic(), Integer.parseInt(m.getValue()[1]));
                 //data.getNode().setStyle("-fx-bar-fill: " + "limegreen"); funktioniert iwie nicht
                 series1.getData().add(data);
             }
@@ -81,7 +81,7 @@ import java.util.Map;
             series2.setName("correct");
 
             for(Map.Entry<Topic, String[]> f: topicResult.entrySet()){
-                series2.getData().add(new XYChart.Data(f.getKey().getTopic(),  Double.parseDouble(f.getValue()[0])));
+                series2.getData().add(new XYChart.Data(f.getKey().getTopic(),  Integer.parseInt(f.getValue()[0])));
             }
 
             barChart.getData().addAll(series1, series2);
